# src/docs/api.yaml
openapi: 3.0.0
paths:
...
  /edzes/{id}/gyakorlat/{userId}:
    post:
      description: >-
        Hozzáad egy gyakorlatot egy meglévő edzéshez és rögzíti a felhasználó
        történetében
      operationId: addGyakorlatToEdzes
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEdzesGyakorlatDto'
      responses:
        '201':
          description: A gyakorlat sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés vagy a gyakorlat nem található
      security:
        - access-token: []
      summary: Gyakorlat hozzáadása edzéshez
      tags:
        - Edzes
  /edzes/{id}/gyakorlat/{gyakorlatId}/{userId}:
    delete:
      description: Törli a gyakorlatot az edzésből és a felhasználó történetéből
      operationId: deleteGyakorlatFromEdzes
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Gyakorlat törlése edzésből
      tags:
        - Edzes
...
  /gyakorlat:
    post:
      operationId: create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGyakorlatDto'
      responses:
        '201':
          description: A gyakorlat sikeresen létrehozva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gyakorlat'
        '400':
          description: Hibás kérés - Érvénytelen adatok vagy izomcsoport azonosítók
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - access-token: []
      summary: Új gyakorlat létrehozása
      tags:
        - Gyakorlat
    get:
      description: >-
        Lapozott lista a gyakorlatokról alapvető információkkal (azonosító, név,
        eszköz és izomcsoportok)
      operationId: findAll
      parameters:
        - name: page
          required: false
          in: query
          description: Oldalszám
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Elemek száma oldalanként
          schema:
            minimum: 1
            maximum: 100
            default: 10
            type: number
        - name: nev
          required: false
          in: query
          description: Szűrés a gyakorlat nevére
          schema:
            type: string
        - name: izomcsoportId
          required: false
          in: query
          description: Szűrés a fő izomcsoport azonosítójára
          schema:
            type: number
        - name: izomcsoportok
          required: false
          in: query
          description: Szűrés izomcsoportokra
          schema:
            type: array
            items:
              type: number
        - name: eszkoz
          required: false
          in: query
          description: Szűrés eszközökre
          schema:
            type: string
      responses:
        '200':
          description: Lapozott gyakorlat lista
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GyakorlatokResponseDto'
        '400':
          description: Hibás kérés - Érvénytelen lekérdezési paraméterek
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - access-token: []
      summary: Összes gyakorlat lekérése
      tags:
        - Gyakorlat
  /gyakorlat/{id}:
    get:
      operationId: findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: A kért gyakorlat adatai
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gyakorlat'
        '404':
          description: A gyakorlat nem található
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GyakorlatNotFoundDto'
      security:
        - access-token: []
      summary: Gyakorlat lekérése azonosító alapján
      tags:
        - Gyakorlat
    patch:
      operationId: update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGyakorlatDto'
      responses:
        '200':
          description: A gyakorlat sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gyakorlat'
        '400':
          description: Hibás kérés - Érvénytelen adatok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: A gyakorlat nem található
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GyakorlatNotFoundDto'
      security:
        - access-token: []
      summary: Gyakorlat módosítása azonosító alapján
      tags:
        - Gyakorlat
    delete:
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: A gyakorlat sikeresen törölve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDto'
        '404':
          description: A gyakorlat nem található
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GyakorlatNotFoundDto'
      security:
        - access-token: []
      summary: Gyakorlat törlése azonosító alapján
      tags:
        - Gyakorlat
  /izomcsoport:
    get:
      description: Lista az összes izomcsoportról
      operationId: findAll
      parameters: []
      responses:
        '200':
          description: Izomcsoport lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Izomcsoport'
        '400':
          description: Hibás kérés - Érvénytelen lekérdezési paraméterek
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - access-token: []
      summary: Összes Izom lekérése
      tags:
        - Izomcsoport
  /user-gyakorlat:
    post:
      description: Létrehoz egy új kapcsolatot a felhasználó és egy gyakorlat között
      operationId: create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserGyakorlatDto'
      responses:
        '201':
          description: A kapcsolat sikeresen létrehozva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlat'
        '400':
          description: A gyakorlat már hozzá van rendelve a felhasználóhoz
        '404':
          description: A gyakorlat vagy a felhasználó nem található
      security:
        - access-token: []
      summary: Új user-gyakorlat kapcsolat létrehozása
      tags:
        - User Gyakorlat
  /user-gyakorlat/history:
    post:
      description: Rögzít egy új edzés eredményt (súly és ismétlésszám) egy gyakorlathoz
      operationId: createHistory
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserGyakorlatHistoryDto'
      responses:
        '201':
          description: A history bejegyzés sikeresen létrehozva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlat'
      security:
        - access-token: []
      summary: Új edzés history bejegyzés létrehozása
      tags:
        - User Gyakorlat
  /user-gyakorlat/user/{userId}:
    get:
      description: Lekéri az összes gyakorlatot és azok történetét egy adott felhasználóhoz
      operationId: getUserGyakorlatok
      parameters:
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: Oldalszám
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Elemek száma oldalanként
          schema:
            minimum: 1
            maximum: 100
            default: 10
            type: number
        - name: isRecord
          required: false
          in: query
          description: ' Ez a tulajdonság arra szolgál, hogy meghatározza, a rekord boolean értékként'
          schema:
            default: false
            type: boolean
        - name: search
          required: false
          in: query
          description: Keresési feltétel gyakorlat neve alapján
          schema:
            type: string
      responses:
        '200':
          description: A gyakorlatok sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlatokResponseDto'
      security:
        - access-token: []
      summary: Felhasználó összes gyakorlatának lekérése
      tags:
        - User Gyakorlat
  /user-gyakorlat/{userId}/{gyakorlatId}:
    get:
      description: >-
        Lekéri egy konkrét gyakorlat részletes adatait és történetét egy
        felhasználóhoz
      operationId: getUserGyakorlatDetails
      parameters:
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
      responses:
        '200':
          description: A gyakorlat adatai sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlat'
        '404':
          description: A gyakorlat nem található a felhasználónál
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlatNotFoundDto'
      security:
        - access-token: []
      summary: Egy konkrét user-gyakorlat részletes adatai
      tags:
        - User Gyakorlat
    delete:
      description: Törli a kapcsolatot egy felhasználó és egy gyakorlat között
      operationId: deleteUserGyakorlat
      parameters:
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
      responses:
        '200':
          description: A gyakorlat sikeresen törölve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlat'
        '404':
          description: A gyakorlat nem található a felhasználónál
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlatNotFoundDto'
      security:
        - access-token: []
      summary: Felhasználó gyakorlatának törlése
      tags:
        - User Gyakorlat
info:
  title: RepVault
  description: |2-

          RepVault API
      
          A védett végpontok eléréséhez Bearer típusú JWT tokent kell használnia.
          Az authentikáció fül alatt lévő /login végponton keresztül tud tokenhez jutni. Adja meg a felhasználónevét és jelszavát,
          illetve ha még nincs felhasználója, regisztráljon az /register végponton keresztül és aztán jelentkezzen be. 
          Ezt követően kattintson az "Authorize" gombra a Swagger felületén, majd másolja be a tokenjét.
      
          Miután megadta a tokent, a Swagger automatikusan hozzáadja azt az "Authorization" fejléchez, 
          így Ön tesztelheti a védett végpontokat.
        
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:8000
    description: Development Server
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    EdzesGyakorlatSet:
      type: object
      properties:
        id:
          type: number
          description: A szett azonosítója
          example: 1
        set_szam:
          type: number
          description: A szett sorszáma
          example: 1
        weight:
          type: number
          description: A gyakorlat súlya
          example: 100.5
        reps:
          type: number
          description: Az ismétlések száma
          example: 12
      required:
        - id
        - set_szam
        - weight
        - reps
...
    EdzesGyakorlat:
      type: object
      properties:
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
        gyakorlat:
          type: object
          description: A gyakorlat neve
          example: Melledzés
        szettek:
          description: A gyakorlat szettjei
          type: array
          items:
            $ref: '#/components/schemas/EdzesGyakorlatSet'
        total_sets:
          type: number
          description: A szettek száma
          example: 12
        previous_history:
          description: Előző gyakorlat előzmények
          type: array
          items:
            $ref: '#/components/schemas/UserGyakorlatHistory'
      required:
        - gyakorlat_id
        - gyakorlat
        - szettek
        - total_sets
        - previous_history
    Edzes:
      type: object
      properties:
        edzes_id:
          type: number
          description: Az edzés azonosítója
          example: 1
        edzes_neve:
          type: string
          description: Az edzés neve
          example: Melledzés
        datum:
          format: date-time
          type: string
          description: Az edzés dátuma
          example: '2024-03-15T10:00:00.000Z'
        ido:
          type: number
          description: Az edzés időtartama (percben)
          example: 60
        isFinalized:
          type: boolean
          description: Igaz ha az edzés befejeződött
          example: false
        isTemplate:
          type: boolean
          description: Igaz ha az edzés sablon
          example: false
        isFavorite:
          type: boolean
          description: Kedvenc edzés
          example: false
        gyakorlatok:
          description: Az edzéshez tartozó gyakorlatok
          type: array
          items:
            $ref: '#/components/schemas/EdzesGyakorlat'
      required:
        - edzes_id
        - edzes_neve
        - datum
        - ido
        - isFinalized
        - isTemplate
        - isFavorite
        - gyakorlatok
...
    AddEdzesGyakorlatDto:
      type: object
      properties:
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
      required:
        - gyakorlat_id
...
    CreateGyakorlatDto:
      type: object
      properties:
        gyakorlat_neve:
          type: string
          example: Fekvenyomás
          description: A gyakorlat megnevezése
        eszkoz:
          type: string
          example: Rúd
          description: A gyakorlathoz szükséges eszköz
        gyakorlat_leiras:
          type: string
          example: Feküdj a padra, engedd le a rudat a mellkasodhoz, majd nyomd fel
          description: A gyakorlat részletes leírása
        fo_izomcsoport:
          type: number
          example: 1
          description: A fő izomcsoport azonosítója
        user_id:
          type: number
          example: 1
          description: A felhasználó azonosítója
          nullable: true
        izomcsoportok:
          example:
            - 1
            - 2
            - 3
          description: Az érintett izomcsoportok azonosítói
          type: array
          items:
            type: number
      required:
        - gyakorlat_neve
        - izomcsoportok
    Gyakorlat:
      type: object
      properties:
        gyakorlat_id:
          type: number
          example: 1
          description: A gyakorlat azonosítója
        gyakorlat_neve:
          type: string
          example: Bench Press
          description: A gyakorlat neve
        eszkoz:
          type: string
          example: Barbell
          description: Eszköz amely szükséges a gyakorlat elvégzéséhez
        gyakorlat_leiras:
          type: string
          example: Feküdj a padra, engedd le a rudat a mellkasodhoz, majd nyomd fel
          description: A gyakorlat leírása
        fo_izomcsoport:
          type: number
          example: 1
          description: Fő izomcsoport azonosítója
        user_id:
          type: number
          example: 1
          description: A felhasználó azonosítója
        izomcsoportok:
          example:
            - 1
            - 2
            - 3
          description: Az érintett izomcsoportok azonosítói
          type: array
          items:
            type: number
      required:
        - gyakorlat_id
        - gyakorlat_neve
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
          description: HTTP státusz kód
        message:
          type: string
          example: Hibás kérés
          description: Hiba üzenet
        error:
          type: string
          example: Bad Request
          description: Hiba típusa
      required:
        - statusCode
        - message
        - error
    GyakorlatListItemDto:
      type: object
      properties:
        gyakorlat_id:
          type: number
          example: 1
          description: A gyakorlat egyedi azonosítója
        gyakorlat_neve:
          type: string
          example: Fekvenyomás
          description: A gyakorlat megnevezése
        eszkoz:
          type: string
          example: Rúd
          description: A gyakorlathoz szükséges eszköz
        fo_izomcsoport:
          type: number
          example: 1
          description: A fő izomcsoport azonosítója
        izomcsoportok:
          example:
            - 1
            - 2
            - 3
          description: Az érintett izomcsoportok azonosítói
          type: array
          items:
            type: number
      required:
        - gyakorlat_id
        - gyakorlat_neve
        - izomcsoportok
    GyakorlatokResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            description: Gyakorlatok listája
            type: array
            items:
              $ref: '#/components/schemas/GyakorlatListItemDto'
        meta:
          description: Lapozási metaadatok
          allOf:
            - $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - items
        - meta
    GyakorlatNotFoundDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
          description: HTTP státusz kód
        message:
          type: string
          example: 'Gyakorlat nem található az ID: 1 alapján'
          description: Hibaüzenet
        error:
          type: string
          example: Not Found
          description: Hiba típusa
      required:
        - statusCode
        - message
        - error
    UpdateGyakorlatDto:
      type: object
      properties:
        gyakorlat_neve:
          type: string
          example: Fekvenyomás
          description: A gyakorlat megnevezése
        eszkoz:
          type: string
          example: Rúd
          description: A gyakorlathoz szükséges eszköz
        gyakorlat_leiras:
          type: string
          example: Feküdj a padra, engedd le a rudat a mellkasodhoz, majd nyomd fel
          description: A gyakorlat részletes leírása
        fo_izomcsoport:
          type: number
          example: 1
          description: A fő izomcsoport azonosítója
        user_id:
          type: number
          example: 1
          description: A felhasználó azonosítója
          nullable: true
        izomcsoportok:
          example:
            - 1
            - 2
            - 3
          description: Az érintett izomcsoportok azonosítói
          type: array
          items:
            type: number
    SuccessResponseDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 200
          description: HTTP státusz kód
        message:
          type: string
          example: A művelet sikeres
          description: Sikeres művelet üzenete
      required:
        - statusCode
        - message
    Izomcsoport:
      type: object
      properties:
        izomcsoport_id:
          type: number
          example: '1'
          description: Az izomcsoport azonosítója
        izomcsoport_neve:
          type: string
          example: hasizom
          description: Az izomcsoport neve
      required:
        - izomcsoport_id
        - izomcsoport_neve
    CreateUserGyakorlatDto:
      type: object
      properties:
        user_id:
          type: number
          description: A felhasználó azonosítója
          example: 1
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
        personal_best:
          type: number
          description: Személyes rekord súly
          example: 100.5
        last_weight:
          type: number
          description: Utoljára használt súly
          example: 90.5
        last_reps:
          type: number
          description: Utolsó ismétlésszám
          example: 12
      required:
        - user_id
        - gyakorlat_id
    UserGyakorlat:
      type: object
      properties:
        user_id:
          type: number
          description: A felhasználó azonosítója
          example: 1
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
        personal_best:
          type: number
          description: Személyes rekord súly
          example: 100.5
        last_weight:
          type: number
          description: Utoljára használt súly
          example: 90.5
        last_reps:
          type: number
          description: Utolsó ismétlésszám
          example: 12
        total_sets:
          type: number
          description: Összes elvégzett szett
          example: 50
        gyakorlat:
          type: object
          description: A gyakorlat adatai
        history:
          description: Edzés történet
          type: array
          items:
            type: object
      required:
        - user_id
        - gyakorlat_id
        - total_sets
    CreateUserGyakorlatHistoryDto:
      type: object
      properties:
        user_id:
          type: number
          description: A felhasználó azonosítója
          example: 1
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
        weight:
          type: number
          description: Az edzésen használt súly
          example: 100.5
        reps:
          type: number
          description: Az elvégzett ismétlések száma
          example: 12
        edzes_id:
          type: number
          description: Az edzés azonosítója
          example: 1
      required:
        - user_id
        - gyakorlat_id
        - weight
        - reps
        - edzes_id
    UserGyakorlatokResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            description: User-Gyakorlat kapcsolatok listája
            type: array
            items:
              $ref: '#/components/schemas/UserGyakorlat'
        meta:
          description: Lapozási metaadatok
          allOf:
            - $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - items
        - meta
    UserGyakorlatNotFoundDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
          description: HTTP státusz kód
        message:
          type: string
          example: 'Gyakorlat nem található az ID: 1 alapján'
          description: Hibaüzenet
        error:
          type: string
          example: Not Found
          description: Hiba típusa
      required:
        - statusCode
        - message
        - error